Introduce a single atom for the table state
Because seems more common to have a single atom. The first time I used
the 2 distinct atoms (forks and eatean) I was facing some difficulties
with the assoc-in/update-in functions and I previosly decided to use 2
different atoms. Now I stepped back and I faced my lack of knowledge.

There is just a single function which modify the state of the forks:
"forks-update!". This commit also rename some of the functions for
consistency.